requirements.txt
# Core dependencies
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2

# Web framework
flask==3.0.0
plotly==5.18.0

# Environment management
python-dotenv==1.0.0

# Development (optional)
pytest==7.4.3
black==23.12.1



